dispose all for many many game challenges to get setState working, https://github.com/phetsims/build-an-atom/issues/156

+++ js/view/PeriodicTableNode.js
@@ -104,15 +104,17 @@ define( function( require ) {
 
     // @private - unlink from Properties
     this.disposePeriodicTableNode = function() {
-      numberAtom.protonCountProperty.unlink( updateHighlightedCell );
+      numberAtom.protonCountProperty.hasListener( updateHighlightedCell ) && numberAtom.protonCountProperty.unlink( updateHighlightedCell );
     };
   }
 
   shred.register( 'PeriodicTableNode', PeriodicTableNode );
   // Inherit from Node.
   return inherit( Node, PeriodicTableNode, {
+
     // @public
     dispose: function() {
+      this.children.forEach( function( node ) { node.dispose();} );
       this.disposePeriodicTableNode();
       Node.prototype.dispose.call( this );
     }
